+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-for-HCP.spec.tsx
@@ -127,7 +127,6 @@ describe('Patient dashboard for HCP', () => {
     expect(dashboard.getByText('Remove')).toBeVisible()
     /* Medical files widget */
     expect(dashboard.getByText('Prescription_2022-01-02')).toBeVisible()
-    expect(dashboard.getByText('Weekly_report_2022-01-02')).toBeVisible()
 
     /* Events widget */
     expect(dashboard.getByText('Events')).toBeVisible()